refactor: Point 패키지 변경
