+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Dto.java
@@ -1,4 +1,4 @@
-package com.loopers.interfaces.api.member.dto;
+package com.loopers.interfaces.api.point;
 
 import java.math.BigDecimal;
